Add an optional token request timeout parameter
The parameter `token_request_timeout` defines how many seconds to wait for the
server to send data before giving up, as a float, or a (connect timeout, read
timeout) tuple. Defaults to None.

By passing float, the timeout value will be applied to both the connect and the
read timeouts. Specify a tuple if you would like to set the values separately.

More detail:
http://docs.python-requests.org/en/master/user/advanced/#timeouts